pyproject.toml
[project]
name = "apprise-api-headless"
description = "Minimalistic version of Apprise API"
readme = "README.md"
license-files = ["LICENSE.md"]
dynamic = ["version"]
requires-python = ">=3.12"
dependencies = [
    "apprise==1.9.5",
    "fastapi[standard]==0.119.1",
    "prometheus-fastapi-instrumentator==7.1.0",
    "pydantic-settings==2.11.0",
    "uvicorn==0.38.0",
]

[dependency-groups]
dev = [
    "httpx==0.28.1",
    "pytest==8.4.2",
    "pytest-httpserver==1.1.3",
    "ruff==0.14.1",
]

[tool.setuptools]
packages = ["apprise_api_headless"]

[tool.pytest.ini_options]
pythonpath = ["."]
log_cli = true
log_cli_level = "DEBUG"
